ARTICLE-SEARCH: see description
the fetch on the page.js is set to revalidate every 600s, however because when the api is fetched, the api is revalidated every 3600s, so it will stale, causing page.js to update the page cache (that revalidate every 600s) to using the stale cache and therefore not attempt to refetch from api again. To fix this, only set cache at the top most level, and keep all data source dynamic.